requirements.txt
fastapi
pymongo
uvicorn
python-dotenv
mangum
